test: move resolve URL context/request tests into file-url-options, add both-URL case
Fold the standalone test/file-url-context.test.js into the existing
test/file-url-options.test.js, and add a case where both the context and
request are URL instances, asserting they combine like new URL(request,
context).
